Demonstrate how to interface Boilermake with an external build system
This contains a demonstration of how a Boilermake-based build system
can be interfaced with an external project that uses some other build
system. In this example, the external project's build system really is
another Boilermake-based build system, but the same techniques can be
used for any other kind of external build system.